requirements.txt
absl-py==2.2.2
altair==5.5.0
argcomplete==3.2.3
astunparse==1.6.3
attrs==25.3.0
awsebcli==3.20.10
blinker==1.9.0
botocore==1.31.85
cachetools==5.5.2
cement==2.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.72.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
narwhals==1.36.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pathspec==0.10.1
pillow==11.2.1
pipx==1.5.0
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==306
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
rich==14.0.0
rpds-py==0.24.0
semantic-version==2.8.5
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
userpath==1.9.2
watchdog==6.0.0
wcwidth==0.1.9
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
